requirements.txt
requests>=2.31
python-dotenv>=1.0
google-generativeai>=0.8
python-dateutil>=2.8
